requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0
feedparser>=6.0.0
anthropic>=0.40.0
authlib>=1.3.0
httpx>=0.25.0
python-dotenv>=1.0.0
pydantic-settings>=2.1.0
itsdangerous>=2.0.0
crawl4ai>=0.8.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth>=2.23.0
pytest>=7.0.0
pytest-mock>=3.10.0
pytest-asyncio>=0.21.0
